+++ packages/util/src/async.ts
@@ -13,3 +13,26 @@ export function isPromise(value: any): value is Promise<any> {
 export function isThenable(value: any): value is PromiseLike<any> {
     return typeof value === 'object' && typeof value.then === 'function';
 }
+
+/**
+ * Waits for the callback to return a true-ish value which resolves the promise.
+ * @param callback Callback function
+ * @param ms Maximum time to wait after which the call 
+ * @param pollInterval The interval at which to execute the callback
+ * @returns Promise
+ */
+export async function poll<T>(callback: () => T, ms: number, pollInterval: number = 50): Promise<T> {
+    const expiryTime = new Date().getTime() + ms;
+    return new Promise<T>((resolve, reject) => {
+        const timer = setInterval(() => {
+            const value = callback();
+            if (value) {
+                clearInterval(timer);
+                resolve(value);
+            } else if (expiryTime < new Date().getTime()) {
+                clearInterval(timer);
+                reject(new Error(`Callback not resolved after waiting for ${ms}`));
+            }
+        }, pollInterval);
+    });
+}