+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog Vlocity/Salesforce Integration for VSCode
 
+## Version 0.16.20 - 2022-02-21
+ - Fix deployment packages get saved before deployment
+ - Fix context actions not visible for related files of datapacks and metadata
+ - Fix changing log viewer visibility doesn't clear logs causing visibility switch to only affect new logs
+
 ## Version 0.16.19 - 2021-29-11
  - Fix LWC refactor doesn't read dirty files correctly causing unintended effect while refactoring of dirty LWC source files